+++ app/jobs/get_game_items_job.rb
@@ -1,0 +1,8 @@
+# Get game items info
+class GetGameItemsJob < ApplicationJob
+  queue_as :default
+
+  def perform
+    GetGameItemsForBankCells.new.call
+  end
+end