+++ db/migrate/20200109173705_add_game_item_to_equipment.rb
@@ -1,0 +1,6 @@
+class AddGameItemToEquipment < ActiveRecord::Migration[5.2]
+  def change
+    add_column :equipment, :game_item_id, :integer
+    add_index :equipment, :game_item_id
+  end
+end