+++ frontend/packs/character_equipment_import.jsx
@@ -1,0 +1,4 @@
+import CharacterEquipmentImport from "components_react/character_equipment_import/character_equipment_import";
+import WebpackerReact from "webpacker-react";
+
+WebpackerReact.setup({CharacterEquipmentImport});